render proposals descritption
